% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultanalyzer.R
\name{resultanalyzer}
\alias{resultanalyzer}
\title{Create result analzer for specific pollutant}
\usage{
resultanalyzer(name)
}
\arguments{
\item{name}{numeric variable}
}
\description{
Create result analzer for specific pollutant
}
