% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{Create time seriers analysics of all numeric variables}
\usage{
timeseries(name)
}
\arguments{
\item{name}{numeric variable}
}
\description{
Create time seriers analysics of all numeric variables
}
